+++ src/promptflow/promptflow/_internal/__init__.py
@@ -41,6 +41,7 @@ from promptflow._core.tools_manager import (
     register_apis,
     register_builtins,
     register_connections,
+    retrieve_tool_func_result,
 )
 from promptflow._core.tracer import Tracer
 from promptflow._sdk._constants import LOCAL_MGMT_DB_PATH
@@ -69,6 +70,7 @@ from promptflow._utils.exception_utils import (
     RootErrorCode,
     infer_error_code_from_class,
 )
+from promptflow._utils.execution_utils import handle_line_failures
 from promptflow._utils.feature_utils import Feature, FeatureState, get_feature_list
 from promptflow._utils.logger_utils import (
     DATETIME_FORMAT,
@@ -99,6 +101,7 @@ from promptflow._utils.utils import (
     transpose,
 )
 from promptflow._version import VERSION
+from promptflow.batch._batch_inputs_processor import apply_inputs_mapping
 from promptflow.executor._errors import InputNotFound
 from promptflow.executor._tool_invoker import DefaultToolInvoker
 from promptflow.storage._cache_storage import LocalCacheStorage