+++ src/promptflow/promptflow/_sdk/_utils.py
@@ -8,7 +8,9 @@ import json
 import logging
 import multiprocessing
 import os
+import platform
 import re
+import stat
 import shutil
 import sys
 import tempfile
@@ -30,9 +32,9 @@ from marshmallow import ValidationError
 from ruamel.yaml import YAML
 
 import promptflow
-from promptflow._constants import EXTENSION_UA
+from promptflow._constants import EXTENSION_UA, PF_NO_INTERACTIVE_LOGIN
 from promptflow._core.tool_meta_generator import generate_tool_meta_dict_by_file
-from promptflow._core.tools_manager import gen_dynamic_list
+from promptflow._core.tools_manager import gen_dynamic_list, retrieve_tool_func_result
 from promptflow._sdk._constants import (
     DAG_FILE_NAME,
     DEFAULT_ENCODING,
@@ -532,6 +534,13 @@ def get_promptflow_sdk_version() -> str:
         return "0.0.1"
 
 
+def print_pf_version():
+    print("promptflow\t\t\t {}".format(get_promptflow_sdk_version()))
+    print()
+    print("Executable '{}'".format(os.path.abspath(sys.executable)))
+    print('Python ({}) {}'.format(platform.system(), sys.version))
+
+
 class PromptflowIgnoreFile(IgnoreFile):
 
     # TODO add more files to this list.
@@ -622,6 +631,31 @@ def _generate_tool_meta(
         error_message = "Generate meta failed, detail error(s):\n" + json.dumps(errors, indent=4)
         raise GenerateFlowToolsJsonError(error_message)
     return res
+
+
+def _retrieve_tool_func_result(func_call_scenario: str, function_config: Dict):
+    """Retrieve tool func result according to func_call_scenario.
+
+    :param func_call_scenario: function call scenario
+    :param function_config: function config in tool meta. Should contain'func_path' and 'func_kwargs'.
+    :return: func call result according to func_call_scenario.
+    """
+    func_path = function_config.get("func_path", "")
+    func_kwargs = function_config.get("func_kwargs", {})
+    # May call azure control plane api in the custom function to list Azure resources.
+    # which may need Azure workspace triple.
+    # TODO: move this method to a common place.
+    from promptflow._cli._utils import get_workspace_triad_from_local
+
+    workspace_triad = get_workspace_triad_from_local()
+    if workspace_triad.subscription_id and workspace_triad.resource_group_name and workspace_triad.workspace_name:
+        result = retrieve_tool_func_result(func_call_scenario, func_path, func_kwargs, workspace_triad._asdict())
+    # if no workspace triple available, just skip.
+    else:
+        result = retrieve_tool_func_result(func_call_scenario, func_path, func_kwargs)
+
+    result_with_log = {"result": result, "logs": {}}
+    return result_with_log
 
 
 def _gen_dynamic_list(function_config: Dict) -> List:
@@ -895,3 +929,35 @@ def dump_flow_result(flow_folder, prefix, flow_result=None, node_result=None):
     if output:
         with open(dump_folder / f"{prefix}.output.json", "w", encoding=DEFAULT_ENCODING) as f:
             json.dump(output, f, indent=2, ensure_ascii=False)
+
+
+def read_write_by_user():
+    return stat.S_IRUSR | stat.S_IWUSR
+
+
+def remove_empty_element_from_dict(obj: dict) -> dict:
+    """Remove empty element from dict, e.g. {"a": 1, "b": {}} -> {"a": 1}"""
+    new_dict = {}
+    for key, value in obj.items():
+        if isinstance(value, dict):
+            value = remove_empty_element_from_dict(value)
+        if value is not None:
+            new_dict[key] = value
+    return new_dict
+
+
+def is_github_codespaces():
+    # Ref:
+    # https://docs.github.com/en/codespaces/developing-in-a-codespace/default-environment-variables-for-your-codespace
+    return os.environ.get("CODESPACES", None) == "true"
+
+
+def interactive_credential_disabled():
+    return os.environ.get(PF_NO_INTERACTIVE_LOGIN, "false").lower() == "true"
+
+
+def is_from_cli():
+    from promptflow._cli._user_agent import USER_AGENT as CLI_UA
+    from promptflow._core.operation_context import OperationContext
+
+    return CLI_UA in OperationContext.get_instance().get_user_agent()