+++ src/promptflow/promptflow/azure/_pf_client.py
@@ -8,7 +8,6 @@ from typing import Dict, List, Optional, Union
 
 from azure.ai.ml import MLClient
 from azure.core.credentials import TokenCredential
-from pandas import DataFrame
 
 from promptflow._sdk._constants import MAX_SHOW_DETAILS_RESULTS
 from promptflow._sdk._errors import RunOperationParameterError
@@ -59,7 +58,10 @@ class PFClient:
         )
         workspace = self._ml_client.workspaces.get(name=self._ml_client._operation_scope.workspace_name)
         self._service_caller = _FlowServiceCallerFactory.get_instance(
-            workspace=workspace, credential=self._ml_client._credential, **kwargs
+            workspace=workspace,
+            credential=self._ml_client._credential,
+            operation_scope=self._ml_client._operation_scope,
+            **kwargs,
         )
         self._flows = FlowOperations(
             operation_scope=self._ml_client._operation_scope,
@@ -67,6 +69,7 @@ class PFClient:
             all_operations=self._ml_client._operation_container,
             credential=self._ml_client._credential,
             service_caller=self._service_caller,
+            workspace=workspace,
             **kwargs,
         )
         self._runs = RunOperations(
@@ -76,6 +79,7 @@ class PFClient:
             credential=self._ml_client._credential,
             flow_operations=self._flows,
             service_caller=self._service_caller,
+            workspace=workspace,
             **kwargs,
         )
         self._connections = ConnectionOperations(
@@ -270,7 +274,7 @@ class PFClient:
 
     def get_details(
         self, run: Union[str, Run], max_results: int = MAX_SHOW_DETAILS_RESULTS, all_results: bool = False
-    ) -> DataFrame:
+    ) -> "DataFrame":
         """Get the details from the run including inputs and outputs.
 
         .. note::