+++ src/promptflow/CHANGELOG.md
@@ -1,6 +1,23 @@
 # Release History
 
-## 0.1.0b9 (Upcoming)
+## 1.1.0 (Upcoming)
+
+### Features Added
+- Add `pfazure flow show/list` to show or list flows from Azure AI.
+- Display node status in run visualize page graph view.
+- Add support for image input and output in prompt flow.
+- [SDK/CLI] SDK/CLI will collect telemetry by default, user can use `pf config set telemetry.enabled=false` to opt out.
+
+### Bugs Fixed
+
+- [SDK/CLI] Fix node test with dict node input will raise "Required input(s) missing".
+
+### Improvements
+- Force 'az login' if using azureml connection provider in cli command.
+- Add env variable 'PF_NO_INTERACTIVE_LOGIN' to disable interactive login if using azureml connection provider in promptflow sdk.
+- Improved CLI invoke time.
+
+## 1.0.0 (2023.11.09)
 
 ### Features Added
 
@@ -11,6 +28,11 @@
 ### Bugs Fixed
 
 - [SDK/CLI] Keep original format in run output.jsonl.
+- [Executor] Fix the bug that raise an error when an aggregation node references a bypassed node
+
+### Improvements
+
+- [Executor] Set the outputs of the bypassed nodes as None
 
 ## 0.1.0b8 (2023.10.26)
 
@@ -50,7 +72,7 @@
 
 - **pf flow validate**: support validate flow
 - **pf config set**: support set user-level promptflow config.
-  - Support workspace connection provider, usage: `pf config set connection.provider=azureml:/subscriptions/<subscription_id>/resourceGroups/<resource_group>/providers/Microsoft.MachineLearningServices/workspaces/<workspace_name>`
+  - Support workspace connection provider, usage: `pf config set connection.provider=azureml://subscriptions/<subscription_id>/resourceGroups/<resource_group>/providers/Microsoft.MachineLearningServices/workspaces/<workspace_name>`
 - Support override openai connection's model when submitting a flow. For example: `pf run create --flow ./ --data ./data.jsonl --connection llm.model=xxx --column-mapping url='${data.url}'`
 
 ### Bugs Fixed